requirements.txt
torch
torchvision
timm
scikit-learn
pillow
streamlit